% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIsing.R
\name{calcIsing}
\alias{calcIsing}
\title{Calculate Ising Sum Component}
\usage{
calcIsing(Z)
}
\arguments{
\item{Z}{a matrix or array of pixel assignments (-1,1)}
}
\value{
the total summation component of the Ising distribution
}
\description{
Helper function in calculating Ising probabilities.
}
