% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcN.R
\name{calcN}
\alias{calcN}
\title{Calculate Neihbourhood Values}
\usage{
calcN(Z)
}
\arguments{
\item{Z}{a matrix or array of pixel assignments (-1,1)}
}
\value{
A matrix of neighbourhood values
}
\description{
This function takes an image of pixel assignments and calcualtes the
neighourhood value for each pixel, average number of 1's in surrounding
neighbourhood.If pixel value is 0, N_ij=0.
}
