% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAvgIsing.R
\name{cAvgIsing}
\alias{cAvgIsing}
\title{Cumulative Average Ising}
\usage{
cAvgIsing(ising_iter, burn = NA)
}
\arguments{
\item{ising_iter}{the iterations from the ising gibbs sampler}

\item{burn}{specified iterations to remove}
}
\value{
a vector of the proportion of times the pixel is assigned to the ROI
}
\description{
Cumulative Average Ising
}
