% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaBound.R
\name{gaBound}
\alias{gaBound}
\title{Genetic Algorithm}
\usage{
gaBound(
  n_pop = 1000,
  n_select = 500,
  max_iter = 1000,
  converge = 0.99,
  crossover = "uniform",
  mutation = "adjacent",
  numCores = 1
)
}
\arguments{
\item{n_pop}{size of initial population}

\item{n_select}{number of parents selected in each set}

\item{max_iter}{maximum number of iterations}

\item{converge}{convergence level}

\item{crossover}{type of crossover}

\item{mutation}{type of mutation}

\item{numCores}{number of cores to run function in parallel}
}
\value{
the total summation component of the Ising distribution
}
\description{
Run GA
}
