% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzParam.R
\name{hzParam}
\alias{hzParam}
\title{Swendsen Wang for Spins}
\usage{
hzParam(z, lambda, tau0, tau1, sigma0_sq, sigma1_sq)
}
\arguments{
\item{z}{pixel assignments}

\item{lambda}{draws from LIRA}

\item{tau0}{mean bkg (sqrt normal)}

\item{tau1}{mean source (sqrt normal)}

\item{sigma0_sq}{var bkg (sqrt normal)}

\item{sigma1_sq}{var source (sqrt normal)}
}
\value{
H
}
\description{
log value
}
