% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPop.R
\name{initPop}
\alias{initPop}
\title{Genetic Algorithm -  Initial Population}
\usage{
initPop(Navg, n_pop)
}
\arguments{
\item{Navg}{the average pixel-wise neighbourhood statistic}

\item{n_pop}{the size of initial population for the GA}
}
\value{
the total summation component of the Ising distribution
}
\description{
Create the initial population to begin the genetic algorithm.
}
