% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isingGibbs.R
\name{isingGibbs}
\alias{isingGibbs}
\title{Run Ising Gibbs Sampler}
\usage{
isingGibbs(
  lambda,
  G,
  init_iter = 500,
  burn_iter = 50,
  beta_jump = 0.01,
  beta_a = 100,
  beta_b = 1,
  beta_start = 0.1,
  beta_niter = 50,
  tau_mu = 5,
  sigma_df = 10,
  omega_sq = 1,
  init_seed = NA,
  ncores = 1
)
}
\arguments{
\item{lambda}{posterior multiscale counts from running LIRA, one per row}

\item{G}{energy information from Beale exact calculation}

\item{init_iter}{number of iterations on inital run}

\item{burn_iter}{burn in for each iteration}

\item{beta_jump}{parameter}

\item{beta_a}{prior parameter}

\item{beta_b}{prior parameter}

\item{beta_start}{initial value for beta parameter (keep close to 0)}

\item{beta_niter}{number of iterations for beta parameter draw}

\item{tau_mu}{prior parameter}

\item{sigma_df}{prior parameter}

\item{omega_sq}{prior parameter}

\item{init_seed}{set seed to replicate results, NA means no seed set}

\item{ncores}{the number of cores , ''all' to use all cores in machine}
}
\value{
the log posterior componants for the ratio calculation
}
\description{
Run the entire Ising Gibbs Sampler to get posterior distribution of pixel assignments
}
