% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPartition.R
\name{loadPartition}
\alias{loadPartition}
\title{Load Partition Function}
\usage{
loadPartition(n, g_file)
}
\arguments{
\item{n}{size of image}

\item{g_file}{file name of partition function}
}
\value{
data frame containing the energy (U) and corresponding counts of each state (G)
}
\description{
Load the partition function from the file name. Partition function is created
using Mathematic files found on Paul Beale's home page:
https://spot.colorado.edu/~beale/index1.html
}
