% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBound.R
\name{multiBound}
\alias{multiBound}
\title{Multi Bound}
\usage{
multiBound(ggobj, bound_counts)
}
\arguments{
\item{ggobj}{gg plot object}

\item{bound_counts}{bound counts vector}
}
\value{
the total summation component of the Ising distribution
}
\description{
Add a series of shaded boundaries onto the plotSource image
Ignore warnings, hacky fix of image limits
}
