% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDraw.R
\name{paramDraw}
\alias{paramDraw}
\title{Conjugate Prior Draw for sqrt Normal}
\usage{
paramDraw(
  Z,
  lambda_draw,
  tau0,
  tau1,
  sigma0_sq,
  sigma1_sq,
  tau_mu,
  sigma_df,
  omega_sq
)
}
\arguments{
\item{Z}{a matrix or array (pixel assignments, spin states)}

\item{lambda_draw}{single draw from LIRA}

\item{tau0}{mean of bkg (sqrt)}

\item{tau1}{mean of source (sqrt)}

\item{sigma0_sq}{var of bkg (sqrt)}

\item{sigma1_sq}{var of source (sqrt)}

\item{tau_mu}{hyperprior}

\item{sigma_df}{hyperprior}

\item{omega_sq}{hyperprior}
}
\value{
a vector that is the product between all adjacent pairs of Z (no particular order)
}
\description{
Conjugate Prior Draw for sqrt Normal
}
