% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Plot A Grid of Images}
\usage{
plotGrid(
  img_list,
  bound_list = NA,
  bound_size = 1,
  coord.range.x = NA,
  coord.range.y = NA,
  crop.range.x = NA,
  crop.range.y = NA,
  save = FALSE,
  legend_label = "Counts",
  plot_file = "dropCurve_boundary.png",
  colors = brewer.pal(9, "Blues"),
  title = "",
  limits = NULL,
  na.value = "grey50",
  width = 11.6,
  height = 9,
  bound_color = "red",
  line_type = 1,
  bkg_color = "black",
  legend_color = "white",
  axis_color = "white",
  show_legend = TRUE,
  legend_horiz = TRUE,
  legend_text_size = 1,
  n_row = length(img_list),
  n_col = 1,
  barwidth = 0.3,
  barheight = 5 * length(img_list),
  label_margin = c(1, 1, 1, 1),
  col_names = NA,
  row_names = NA
)
}
\arguments{
\item{img_list}{image}

\item{bound_list}{list of boundaries}

\item{bound_size}{width of boundary}

\item{coord.range.x}{coord range x}

\item{coord.range.y}{coordinate range y}

\item{crop.range.x}{crop}

\item{crop.range.y}{crop}

\item{save}{TRUE if save}

\item{legend_label}{legend title}

\item{plot_file}{destination of saved file}

\item{colors}{color scale}

\item{title}{title of image}

\item{limits}{limits of scale}

\item{na.value}{NA value color}

\item{width}{width of saved image}

\item{height}{height of saved image}

\item{bound_color}{color of bound or vector of colors of bounds}

\item{line_type}{shape of line for boundary}

\item{bkg_color}{color of bkg to image}

\item{legend_color}{color of legend text}

\item{axis_color}{color of axis numbers,text}

\item{show_legend}{show legend  on image}

\item{legend_horiz}{orientation of the legend}

\item{legend_text_size}{size of text of the legend}

\item{n_row}{number of rows in grid}

\item{n_col}{number of columns in legend}

\item{barwidth}{legend width}

\item{barheight}{legend heigh}

\item{label_margin}{margins}

\item{col_names}{names of columns}

\item{row_names}{names of rows}
}
\value{
the total summation component of the Ising distribution
}
\description{
Plot the source. Options to include coordinates and boundaries.
}
