% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postRatio.R
\name{postRatio}
\alias{postRatio}
\title{Calculate Log Posterior}
\usage{
postRatio(Z1, Z2, lambda, param)
}
\arguments{
\item{Z1}{vector of matrix representation of top pixel assignments (-1,1)}

\item{Z2}{vector of matrix representation of bottom pixel assignments (-1,1)}

\item{lambda}{posterior draws of LIRA output, one draw per row}

\item{param}{posterior draws of other parameters, one draw per row}
}
\value{
the log posterior componants for the ratio calculation
}
\description{
Calculates componants of the ratio to compare two posterior distriubtions
of the pixel assignments
}
