% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioCheck.R
\name{ratioCheck}
\alias{ratioCheck}
\title{Calculate Final Ration}
\usage{
ratioCheck(log1, log2)
}
\arguments{
\item{log1}{the log posterior componants for pixel assignments 1}

\item{log2}{the log posterior componants for pixel assignments 2}
}
\value{
the weights and exponant components of the ratio
}
\description{
This function makes use of the Brobdingnag for adding exponants.
Even after the adjustment made to alleviate underflow error, there are still
issues calculating the ratio. Adding the probabilities as type brob
reduces underflow error even more.
}
