% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectBase.R
\name{rectBase}
\alias{rectBase}
\title{Rectangle Base for Simulation}
\usage{
rectBase(bkg_param, max_param, n_img, width, height)
}
\arguments{
\item{bkg_param}{value of bkg noise}

\item{max_param}{value of maximum step}

\item{n_img}{size of pixels of image}

\item{width}{width of steps}

\item{height}{height of steps}
}
\description{
Create a rectangle from which to simulate observed images from
}
