% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducePop.R
\name{reproducePop}
\alias{reproducePop}
\title{Genetic Algorithm -  Selection}
\usage{
reproducePop(select_population, crossover = "uniform", mutation = "adjacent")
}
\arguments{
\item{select_population}{population to reproduce}

\item{crossover}{type of crossover}

\item{mutation}{type of mutation}
}
\value{
the total summation component of the Ising distribution
}
\description{
Tournament selection. Randomly select two parents from the inital populatio
with replacement. Add the fittest to the selection population. Do this until
you have a specified number of parents in the selection population.
}
