% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGibbs.R
\name{runGibbs}
\alias{runGibbs}
\title{Run Ising Gibbs Sampler}
\usage{
runGibbs(
  lira,
  G,
  init_iter = 500,
  burn_iter = 50,
  bet.jump = 0.01,
  beta.a = 100,
  beta.b = 1,
  beta.start = 0.1,
  beta.niter = 50,
  tau.mu = 5,
  sigma.df = 10,
  omega.sq = 1,
  init.seed = NA,
  ncores = NA
)
}
\arguments{
\item{lambda}{posterior multiscale counts from running LIRA, one per row}
}
\value{
the log posterior componants for the ratio calculation
}
\description{
Run the Ising Gibbs Sampler to get posterior distribution of pixel assignments
}
