% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNode.R
\name{sampleNode}
\alias{sampleNode}
\title{Sample Nodes}
\usage{
sampleNode(z_node, tau0, tau1, sigma0_sq, sigma1_sq)
}
\arguments{
\item{z_node}{nodes of spin states}

\item{tau0}{mean of bkg (sqrt)}

\item{tau1}{mean of source (sqrt)}

\item{sigma0_sq}{var of bkg (sqrt)}

\item{sigma1_sq}{var of source (sqrt)}
}
\value{
sample new spin states for each cluster
}
\description{
SW step
}
