% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempMH.R
\name{tempMH}
\alias{tempMH}
\title{Draw Temperature Parameter (beta)}
\usage{
tempMH(beta, z, G, jump, n, beta_a, beta_b, iter)
}
\arguments{
\item{beta}{current beta step}

\item{z}{current spin state}

\item{G}{data frame with energy and density for partition}

\item{jump}{size of jump}

\item{n}{size of image}

\item{beta_a}{hyper prior for beta}

\item{beta_b}{hyper prior for beta}

\item{iter}{number of iterations}
}
\value{
log of the partition function given a specific energy state and density
}
\description{
Draws for the temperature parameter (beta) using Metropolis Hastings
}
